fix: swap workspace & managed selector for clarity
